# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.5"
  title: SiB Colombia - Catalogue API
  description: API for SiB Colombia Catalogue. This API is used as a backend for consult the information of the species records reported in the SiB Colombia.
  termsOfService: http://swagger.io/terms/
  contact:
    name: SiB Colombia Developer Team
    url: http://www.sibcolombia.net
    email: oduque@humboldt.org.co
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /api/v1.5
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /records/{id}/full_description/{version}:
    x-swagger-router-controller: fullDescription
    get: 
      description: Returns a FullDescriptionVersion by id_record and number of version
      tags:
        - FullDescription
      operationId: getFullDescription
      parameters:
        - name: id
          type: string
          description: "ID of the record(ficha)"
          in: path
          required: true
        - name: version
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/FullDescriptionVersion"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /records/{id}/full_description:
    x-swagger-router-controller: fullDescription
    post:
      description: Save a new  FullDescriptionVersion that belongs to the record with the id in the route
      tags:
        - FullDescription
      operationId: postFullDescription
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: fullDescriptionVersion
          description: fullDescriptionVersion object
          in: body
          required: true
          schema:
            $ref: '#/definitions/FullDescriptionVersion'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResultResponse"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
  /records/to_review/{id}/full_description:
    x-swagger-router-controller: fullDescription
    get: 
      description: Returns a list of FullDescriptionVersion with state "to_review"
      tags:
        - FullDescription
      operationId: getToReviewFullDescription
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/FullDescriptionVersionList"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /group/{id}:
    x-swagger-router-controller: group
    get: 
      description: Returns a group by id
      tags:
        - group
      operationId: getGroup
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Group"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /group:
    x-swagger-router-controller: group
    post:
      description: Save a new group 
      tags:
        - group
      operationId: postGroup
      parameters:
        - name: group
          description: Group object
          in: body
          required: true
          schema:
            $ref: '#/definitions/Group'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResultResponse"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  FullDescriptionVersionList:
    properties:
      fullDescriptionList:
        type: array
        items:
          $ref: '#/definitions/FullDescriptionVersion'
...
  FullDescriptionVersion:
    required:
      - fullDescription
      - id_user
    properties:
      -id:
        type: string
      id_record: 
        type: string
      created: 
        type: string #****
      id_user:
        type: string
      version:
        type: integer
      state:
        type: string
      element:
        type: string
      fullDescription:
        $ref: '#/definitions/FullDescription'
...
  Agent:
    properties:
      firstName: 
        type: string
      lastName:
        type: string
      organisation:
        type: string
      position:
        type: string
      address:
        type: string
      city:
        type: string
      state:
        type: string
      country: 
        type: string
      postalCode:
        type: string
      phone:
        type: string
      email:
        type: string
      homepage:
        type: string
      personnelDirectory:
        type: string
      personnelIdentifier :
        type: string
...
  FullDescription:
    properties:
      ancillaryData:
        type: array
        items:
          $ref: '#/definitions/AncillaryData'
      fullDescriptionUnstructured:
        type: string
      fullDescriptionAtomized:
        type: array
        items:
          $ref: '#/definitions/FullDescriptionAtomized'   
  FullDescriptionAtomized:
    properties:
      measurementOrFact:
        $ref: '#/definitions/MeasurementOrFact'
      ancillaryData:
        $ref: '#/definitions/AncillaryData'
...
  MeasurementOrFact:
    properties:
      measurementID:
        type: string
      measurementType:
        type: string
      measurementValue:
        type: string
      measurementAccuracy:
        type: string
      measurementUnit:
        type: string
      measurementDeterminedDate: #**
        type: string
      measurementDeterminedBy:
        type: array
        items: 
          type: string
      measurementMethod:
        type: string
      measurementRemarks:
        type: string
      relatedTo:
        type: string
...
  AncillaryData:
    properties:
      dataType:
        type: string 
      mimeType:
        type: string
      agent:
        type: array
        items:
          $ref: '#/definitions/Agent' 
      created:
        type: string
        #pattern: '^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$'
      modified:
        type: string
        #pattern: '^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$'
      title:
        type: string
      license:
        type: string
      rights:
        type: string
      rightsHolder:
        type: string
      bibliographicCitation:
        type: string
      audience:
        type: array
        items:
          type: string
      source :
        type: string
      subject:
        type: array
        items:
          type: string
      description:
        type: string
      mediaURL:
        type: array
        items:
          type: string
      thumbnailURL:
        type: string
      location:
        type: string
      geoPoint:
        type: string
      additionalInformation:
        type: string
      dataObject:
        type: string
      reference: 
        type: array
        items:
          $ref: '#/definitions/Reference'      
  Reference:
    properties:
      profile_id:
        type: string
      group_id:
        type: string
      created:
        type: string
        #pattern: '^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$'
      last_modified:
        type: string
        #pattern: '^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$'
      identifiers:
        type: array
        items:
          type: string
      abstract:
        type: string
      type:
        type: string
      source: 
        type: string
      title:
        type: string
      authors:
        type: array
        items:
          type: string
      year:
        type: string
      volume: 
        type: string
      issue:
        type: string
      pages:
        type: string
      series:
        type: string
      chapter: 
        type: string
      websites:
        type: string
      accessed:
        type: string
      publisher:
        type: string
      address:
        type: string
      city:
        type: string
      edition:
        type: string
      institution:
        type: string
      editors:
        type: array
        items:
          type: string
      keywords:
        type: array
        items:
          type: string
      doi:
        type: string
      isbn:
        type: string
      issn:
        type: string
      link:
        type: string
      recordId:
        type: string
...
  Group:
    required:
      - users
      - admins
    properties:
      users:
        type: array
        items:
          type: string
      admins:
        type: array
        items:
          type: string
      image: 
        type: string
      created: 
        type: string
      description: 
        type: string
      tags:
        type: array
        items:
          type: string
      records: 
        type: string
        items:
          type: string
...
  ResultResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string